Renamed modules as well
