+++ crates/dojo-bindgen/src/plugins/typescript_v2/mod.rs
